Move to microservices (at least partially)
