+++ .rspec
@@ -1,0 +1,4 @@
+--color
+--require spec_helper
+-f d
+--order random
