+++ lib/services/fetch_products.rb
@@ -1,0 +1,7 @@
+module Store
+	class FetchProducts
+    def call 
+      PRODUCTS
+    end
+	end
+end
