+++ lib/services/fetch_warehouse.rb
@@ -1,0 +1,6 @@
+module Store
+  class FetchWarehouse
+    def call(name) 
+      WAREHOUSE.fetch { |w| w.name == name } 
+  end
+end
